FBA: optimizaciones de rendimiento (timers O(1), APU int, PPU render)
Profiling headless: el CPU rinde ~46 MIPS (no es el cuello), pero los ticks de
subsistemas consumian ~77% del frame. Optimizaciones aplicadas:
- TimerController.tick: la rutina recorria el prescaler ciclo-a-ciclo (con
  prescaler=1, usado por el audio Direct Sound, eran ~280k iteraciones/frame por
  timer). Ahora calcula los incrementos por aritmetica O(1) y solo itera por
  overflow real. Medido: +27% en el coste de timers.
- APU: acumuladores de ciclos pasados de double a int (CYCLES_PER_SAMPLE=512
  exacto) -> menos coste en el hot path.
- PPU.renderScanline: init de buffers de composicion con Arrays.fill (intrinseco
  vectorizado) en vez de loop por pixel; computeWindows usa Arrays.fill cuando no
  hay ventanas; camino rapido en el compositor cuando no hay blending/ventana/OBJ
  semitransparente (convierte top-layer directo a ARGB).
- GBAEmulator: tracer.isEnabled() cacheado una vez por frame.

Verificado: 103/103 tests OK; RtCheck 60 FPS estables; escena de Abedul sigue con
color correcto (0 px piel turquesa, 107 px piel). Sin regresiones.

+++ gba-minecraft-mod/src/main/java/com/gbaminecraft/emulator/apu/APU.java
@@ -52,10 +52,11 @@ public class APU {
     private boolean ch4Running = false;
     private int ch4LenTimer = 0; boolean ch4LenEnabled = false;
 
-    // Cycle accumulators
-    private double ch1CycAcc = 0, ch2CycAcc = 0, ch3CycAcc = 0, ch4CycAcc = 0;
-    private double sampleCycAcc = 0;
-    private static final double CYCLES_PER_SAMPLE = (double)CPU_FREQ / SAMPLE_RATE;
+    // Cycle accumulators (integer: CYCLES_PER_SAMPLE is exactly 512, no need for
+    // floating point in the per-instruction hot path).
+    private int ch1CycAcc = 0, ch2CycAcc = 0, ch3CycAcc = 0, ch4CycAcc = 0;
+    private int sampleCycAcc = 0;
+    private static final int CYCLES_PER_SAMPLE = CPU_FREQ / SAMPLE_RATE; // = 512 exactly
 
     // Frame sequencer (512 Hz from timer)
     private int frameSeqTimer = 0;
